Added workarounds for user "error"
Added a piece of code that lowers the user input and the fields of the data frame in case the user decides to input the country uncapitalised or decides to start with a capitalised version of one of the words that I use in my RegEx to find the parameters I am looking for inside the input. 
Moved the nltk downloaders to the data skimmer, so that the program doesn't print to console the update of the keywords every time the program is run. This also presupposes that the user always runs the data skimmer first, which should be the case.